requirements.txt
flask
pandas
numpy
scikit-learn
pyspark
joblib
pytest
scikit-surprise
